Start removal of web UI config interface
As the configuration implementation writes all default fields regardless
if they are present in file it sort of messes up the structure of the
config file.

The need to update the configuration at runtime is small as I see it.
Combined with the process has to be restarted for the changes to
have an impact it not a very valueable interface.

As the trackers are no longer hardcoded it makes no sense keeping and
maintaining the interface.

Thus it is being removed.