Add command line flags for config locations
One flag for specifying where the tracker XML configurations are
located.

One flag for specifying the user configuration file. Since specifying
the path requires parsing of input parameters before parsing config the
config cannot be fetched in the global scope of each module. The API to
the config has either been extended to fetch the data when necessary or
the config is given as a parameter.

Because of this it was convenient to also only notify configured
backends.