New database design
* Use database relations instead of duplicating data.
* Move database queries to db module

Since the query for snatches returns tuples it is a bit more
inconvienient when using the data. But for the greater good is should be
fine. Converting to named tuple or similar seems like a waste.

Convert script included to read the old design from one file and write
the new design to another one.