Cast host and port config values
Tomlkit uses its own types for int and string which causes problem with
the standard library. More specifically the name resolution for URLs
fail.

Have not yet seen any issue with the webui host and port but since it
might be used in the same way somewhere underneath they are casted as
well.