Decrease db_session scope
Pony is not compatible with async/await statements. More specifically,
when the context manager is active await is not possible. The database
sessions are therefore changed to what's essential even though it may
cause several session per event. Multiple session per event is rare
though, it only happens if a release is snatched. Should be fine.