Move shutdown to signal handler
socketio replaces the werkzeug dev webserver when using eventlet and so
the shutdown endpoint does no longer work. Moving the shutdown to a
signal handler, which makes sense anyway, as it's easy to make graceful
shutdown from CLI.

IRC and db threads as well as backend's session are shutdown and closed
properly. The web thread has no natural flag or event for shutdown that
I can find. There is probably some hacky way to send SystemExit
exception to it but might as well just terminate the program.